Rename `builder` module to `redactor`
